header refactoring
