+++ front/components/layouts/Header.tsx
@@ -1,24 +1,13 @@
 import { FC } from 'react';
-import Link from 'next/link';
 import style from '../../styles/navbar.module.css';
 import { useUserStore } from '../../store/store';
 import { useNavStore } from '../../store/store';
-import { useLogOut } from '../../hooks/useLogOut';
+
+import { HomeEl, FavEl, LoginEl, AccountEl, LogoutEl } from './Navbar';
 export const Header: FC = () => {
   const { user } = useUserStore();
   const { show } = useNavStore();
   const toggle = useNavStore((show) => show.toggle);
-  const resetUser = useUserStore((state) => state.resetUser);
-
-  const logOut = () => {
-    useLogOut()
-      .then((res) => {
-        resetUser();
-      })
-      .catch((error) => {
-        console.log(error);
-      });
-  };
 
   return (
     <header className="fixed w-screen backdrop-blur-2xl">
@@ -36,55 +25,28 @@ export const Header: FC = () => {
             {show ? (
               <ul className={show ? style.all : undefined}>
                 <div className=" text-white flex flex-col h-screen items-center justify-center">
-                  <li className="text-lg font-bold py-5">
-                    <Link href="/">Home</Link>
-                  </li>
-                  <li className="text-lg font-bold py-5">
-                    <Link href="/favorite">Favorite</Link>
-                  </li>
+                  <HomeEl />
+                  <FavEl />
                   {user.id == '' ? (
-                    <li className="text-lg font-bold py-5">
-                      <Link href="/account/login">Login/signup</Link>
-                    </li>
+                    <LoginEl />
                   ) : (
                     <div>
-                      <button
-                        className="text-lg font-bold py-5"
-                        onClick={() => logOut()}
-                      >
-                        LogOut
-                      </button>
-                      <li className="text-lg font-bold py-5">
-                        <Link href="/account/edit">edit</Link>
-                      </li>
+                      <LogoutEl />
+                      <AccountEl />
                     </div>
                   )}
                 </div>
               </ul>
             ) : (
               <ul className="md:flex hidden gap-10 justify-end">
-                <li className="text-lg font-bold">
-                  <Link href="/">Home</Link>
-                </li>
-                <li className="text-lg font-bold">
-                  <Link href="/favorite">Favorite</Link>
-                </li>
+                <HomeEl />
+                <FavEl />
                 {user.id == '' ? (
-                  <li className="text-lg font-bold">
-                    <Link href="/account/login">Login/signup</Link>
-                  </li>
+                  <LoginEl />
                 ) : (
                   <div className="flex gap-10">
-                    <button
-                      className="text-lg font-bold "
-                      onClick={() => logOut()}
-                    >
-                      LogOut
-                    </button>
-                    <li className="text-lg font-bold ">
-                      <Link href="/account">account</Link>
-                    </li>
-                    <li className="text-lg font-bold ">{user.name}</li>
+                    <LogoutEl />
+                    <AccountEl />
                   </div>
                 )}
               </ul>
